Refactor: SettingAction 컴포넌트를 상대 경로에서 절대 경로로 변경하여 코드 가독성 향상
