+++ client/src/widgets/settings-modal/components/AllArchiveChatRoomSetting.tsx
@@ -1,4 +1,4 @@
-import SettingAction from "./SettingAction";
+import { SettingAction } from "@/widgets/settings-modal/components";
 
 const AllArchiveChatRoomSetting = () => {
   const handleArchive = () => {
