+++ client/src/widgets/settings-modal/components/index.ts
@@ -2,5 +2,12 @@ import ThemeSelectSetting from "./ThemeSelectSetting";
 import AllDeleteChatRoomSetting from "./AllDeleteChatRoomSetting";
 import AllArchiveChatRoomSetting from "./AllArchiveChatRoomSetting";
 import ArchiveChatRoomManage from "./ArchiveChatRoomManage";
+import SettingAction from "./SettingAction";
 
-export { ThemeSelectSetting, AllDeleteChatRoomSetting, AllArchiveChatRoomSetting, ArchiveChatRoomManage };
+export {
+  ThemeSelectSetting,
+  AllDeleteChatRoomSetting,
+  AllArchiveChatRoomSetting,
+  ArchiveChatRoomManage,
+  SettingAction,
+};
